warnings cleanup

+++ src/CinnamonScreenSaverInhibitor.cpp
@@ -30,7 +30,13 @@ THIS::THIS(std::function<void(Inhibitor*, Inhibit)> inhibitCB,
 
 void THIS::handleSimActivityMsg(DBus::Message* msg, DBus::Message* retmsg) {
 	// We treat this as an inhibit that expires in 5min
-	Inhibit i = {InhibitType::SCREENSAVER, msg->sender(), "SimulateUserActivity", this->mkId(msg->sender()), time(NULL)};
+	Inhibit i = {
+		InhibitType::SCREENSAVER, 
+		msg->sender(), 
+		"SimulateUserActivity", 
+		this->mkId(msg->sender()), 
+		(uint64_t)time(NULL)
+	};
 
 	// Clear any existing simActivity inhibits from this sender	
 	std::vector<InhibitID> eraseIDs;
